silence delete() if the file does not exist; cleans up 'make data' invocation

+++ data/corpus/google-ngrams/fetch.py
@@ -64,9 +64,10 @@ def extract(filename, gzfile):
 				out.write('%s\t%s\n' % (k, d[k]))
 
 def delete(filename):
-	log(filename, 'delete')
 	try:
-		os.remove(filename)
+		if os.path.exists(filename):
+			log(filename, 'delete')
+			os.remove(filename)
 	except:
 		pass
